Finesampling fix galore
"finesamling" at the R ProFit level depends on the "pixel resolution" feature
of libprofit, which is the ability to specify how big a pixel is in image
coordinates, separately for the X and Y axis. So far libprofit was handling
pixel resolutions correctly for the sersic profile, but psf profiles were not
taking them into account at all. This has now been fixed. Moreover, the PSF
image itself can also specify its own horizontal and vertical resolution.

With this support now in place I've updated the profitMakeModel code to
properly set the sample_x/sample_y model attributes depending on the given
finesample parameter. I also added a big comment explaining how things work.

The example plots from the paper seem to be working fine, even when creating
them with finesampling.

+++ src/libprofit/src/convolve.cpp
@@ -82,8 +82,8 @@ double *convolve(double *src, unsigned int src_width, unsigned int src_height,
 
 					src_i = (int)i + (int)k - (int)krn_half_width;
 
-					if( src_i >= 0 && src_i < src_width &&
-					    src_j >= 0 && src_j < src_height ) {
+					if( src_i >= 0 && (unsigned int)src_i < src_width &&
+					    src_j >= 0 && (unsigned int)src_j < src_height ) {
 						pixel +=  *srcPtr2 * *krnPtr;
 					}
